requirements.txt
pg8000>=1.30.0
python-dotenv>=1.0.0